refactor: Move controller types into separate module
