+++ src/lib.rs
@@ -7,11 +7,10 @@ use embedded_io::ReadExactError;
 mod fmt;
 
 pub mod cmd;
+pub mod controller;
 pub mod data;
-pub mod driver;
 pub mod event;
 pub mod param;
-pub mod serial;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
@@ -363,37 +362,4 @@ impl<'a, T: HostToControllerPacket> WriteHci for WithIndicator<'a, T> {
         T::KIND.write_hci_async(&mut writer).await?;
         self.0.write_hci_async(writer).await
     }
-}
-
-pub trait Controller {
-    type Error: embedded_io::Error;
-
-    fn write_acl_data(&self, packet: &data::AclPacket) -> impl Future<Output = Result<(), Self::Error>>;
-    fn write_sync_data(&self, packet: &data::SyncPacket) -> impl Future<Output = Result<(), Self::Error>>;
-    fn write_iso_data(&self, packet: &data::IsoPacket) -> impl Future<Output = Result<(), Self::Error>>;
-
-    fn read<'a>(&self, buf: &'a mut [u8]) -> impl Future<Output = Result<ControllerToHostPacket<'a>, Self::Error>>;
-}
-
-#[derive(Debug)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum CmdError<E> {
-    Param(param::Error),
-    Controller(E),
-}
-
-impl<E> From<param::Error> for CmdError<E> {
-    fn from(e: param::Error) -> Self {
-        Self::Param(e)
-    }
-}
-
-pub trait ControllerCmdSync<C: cmd::SyncCmd + ?Sized>: Controller {
-    /// Note: Some implementations may require [`Controller::read()`] to be polled for this to return.
-    fn exec(&self, cmd: &C) -> impl Future<Output = Result<C::Return, CmdError<Self::Error>>>;
-}
-
-pub trait ControllerCmdAsync<C: cmd::AsyncCmd + ?Sized>: Controller {
-    /// Note: Some implementations may require [`Controller::read()`] to be polled for this to return.
-    fn exec(&self, cmd: &C) -> impl Future<Output = Result<(), CmdError<Self::Error>>>;
 }
